hwmontempsensor: Use isChassisOn() in isDeviceOff()
The temp sensors on PCIe slot power, such as the TMP435s on the Flett
and Bear River cards, can only be accessed when the slot is powered on.
This was being checked for by reading the slot power state property.

However, there are cases where that property doesn't react fast enough
to power offs, such as when a PGood fault brings down the system.  This
would lead to a read failure error log on those sensors because the code
still thought the power was on.

To fix this, also add a check to see if system power is off when
checking to see if device on slot power is off, and then check the PGood
property inside the isChassisOn() function as well as the chassis power
state.  That function is also checked before creating an error log, so
it will also prevent errors from being logged when the PGood is off but
the status hasn't been reflected in the chassis power state yet.

It may be possible to change the isChassisOn() on from 'pgood on and
power state on' to just 'pgood on'.  That would open up a bit of window
when PGood is on before the power state gets set to on that error logs
could now be created.  That doesn't seem like a big deal, but I'm going
to err on the side of caution for now and not mess with it.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I751c5203dc6095e02ec1f0c4c1654551397686af