Revert "hwmontempsensor: Simplify I/O and buffering"
This reverts commit d1c8f4433e981696d06e3c29006affd96624c89e.

We want to revert running hwmontempsensor with io_uring because if it
restarts at just the right time, such as during a power on, it may fail
with io_uring complaining it doesn't have enough memory, eventually
leading to the quiesce state.

However, just turning off the io_uring boost compile flags won't compile
as-is, and fails with errors involving boost::asio::random_access_file.

It appears that random_access_file, which was introduced with the commit
being reverted here, only works with io_uring, or there is just some
other combination of compile flags or includes I don't understand.

The commit was introduced to fix issue openbmc/dbus-sensors/#20, which
is possibly a use-after-free bug that can occur during stress testing.
IBM has never seen this problem in 1020/1030 which didn't have this
commit, and I think that a crash in this case would lead to a
successful restart, unlike the io_uring crash we're trying to fix.

Tested:
hwmontempsensor still works fine on our systems